Add tests for reproduceability, and make sure they succeed.
(1) Words are now generated at random, not from by counter increment
(2) Language-to-string is now sorted
(3) simulate is now using a loop variable that is different from the argument
(4) The random seed for Multiprocess has been moved down

In python versions with a hash seed (3.5, in particular) it is still necessary
to run the tests with PYTHONHASHSEED set, eg. to 0

The test also checks consistency between runs, for which in needs a helper JSON
file. Missing entries in this file, or a completely missing file, are ignored.